+++ src/value/string/index.ts
@@ -3,6 +3,7 @@
 export * from "./canonical-uri.schema.js";
 export * from "./canonical-uriwith-knowledge-time.schema.js";
 export * from "./cleaned-string.schema.js";
+export * from "./email-address.schema.js";
 export * from "./language.schema.js";
 export * from "./relative-uri.schema.js";
 export * from "./telephone-number.schema.js";